Remove miniwin
Event handling code moved to `engine/events.{hpp,cpp}`.
