+++ Source/control.h
@@ -59,6 +59,12 @@ extern SDL_Rect PanBtnPos[8];
 
 void CalculatePanelAreas();
 bool IsChatAvailable();
+
+/**
+ * @brief Moves the mouse to the first attribute "+" button.
+ */
+void FocusOnCharInfo();
+
 /**
  * @brief Check if the UI can cover the game area entierly
  */
