+++ Source/controls/game_controls.cpp
@@ -3,7 +3,6 @@
 #include <cstdint>
 
 #include "controls/controller_motion.h"
-#include "miniwin/misc_msg.h"
 #ifndef USE_SDL1
 #include "controls/devices/game_controller.h"
 #endif
