+++ Source/debug.cpp
@@ -15,13 +15,13 @@
 #include "control.h"
 #include "cursor.h"
 #include "engine/backbuffer_state.hpp"
+#include "engine/events.hpp"
 #include "engine/load_cel.hpp"
 #include "engine/point.hpp"
 #include "error.h"
 #include "inv.h"
 #include "levels/setmaps.h"
 #include "lighting.h"
-#include "miniwin/misc_msg.h"
 #include "monstdat.h"
 #include "monster.h"
 #include "plrmsg.h"
