+++ Source/engine/events.hpp
@@ -1,27 +1,24 @@
-/**
- * @file miniwin/misc_msg.h
- *
- * Contains most of the the demomode specific logic
- */
 #pragma once
 
-#include <SDL.h>
 #include <cstdint>
+
+#include <SDL.h>
+
+#include "engine/point.hpp"
 
 #ifdef USE_SDL1
 #include "utils/sdl2_to_1_2_backports.h"
 #endif
 
-#include "engine/point.hpp"
-
 namespace devilution {
 
 using EventHandler = void (*)(const SDL_Event &event, uint16_t modState);
 
-void SetCursorPos(Point position);
-void FocusOnCharInfo();
+/** @brief The current input handler function */
+extern EventHandler CurrentEventHandler;
 
-void SetMouseButtonEvent(SDL_Event &event, uint32_t type, uint8_t button, Point position);
+EventHandler SetEventHandler(EventHandler NewProc);
+
 bool FetchMessage(SDL_Event *event, uint16_t *modState);
 
 void HandleMessage(const SDL_Event &event, uint16_t modState);
