+++ Source/engine/render/scrollrt.cpp
@@ -25,9 +25,6 @@
 #include "inv.h"
 #include "lighting.h"
 #include "minitext.h"
-#ifdef _DEBUG
-#include "miniwin/misc_msg.h"
-#endif
 #include "missiles.h"
 #include "nthread.h"
 #include "options.h"
