+++ Source/gmenu.cpp
@@ -14,7 +14,6 @@
 #include "engine/load_cel.hpp"
 #include "engine/render/clx_render.hpp"
 #include "engine/render/text_render.hpp"
-#include "miniwin/misc_msg.h"
 #include "options.h"
 #include "stores.h"
 #include "utils/language.h"
