+++ Source/init.h
@@ -5,18 +5,20 @@
  */
 #pragma once
 
-#include "miniwin/misc_msg.h"
 #include "utils/attributes.h"
 #include "utils/stdcompat/optional.hpp"
 
-#ifndef UNPACKED_MPQS
+#ifdef UNPACKED_MPQS
+#include <string>
+#else
 #include "mpq/mpq_reader.hpp"
 #endif
+
+#include <SDL.h>
 
 namespace devilution {
 
 extern bool gbActive;
-extern EventHandler CurrentEventHandler;
 extern DVL_API_FOR_TEST bool gbIsSpawn;
 extern DVL_API_FOR_TEST bool gbIsHellfire;
 extern DVL_API_FOR_TEST bool gbVanilla;
@@ -87,6 +89,5 @@ void LoadLanguageArchive();
 void LoadGameArchives();
 void init_create_window();
 void MainWndProc(const SDL_Event &event);
-EventHandler SetEventHandler(EventHandler NewProc);
 
 } // namespace devilution
