+++ Source/interfac.cpp
@@ -13,6 +13,7 @@
 #include "engine/clx_sprite.hpp"
 #include "engine/demomode.h"
 #include "engine/dx.h"
+#include "engine/events.hpp"
 #include "engine/load_cel.hpp"
 #include "engine/load_clx.hpp"
 #include "engine/load_pcx.hpp"
@@ -21,7 +22,6 @@
 #include "hwcursor.hpp"
 #include "init.h"
 #include "loadsave.h"
-#include "miniwin/misc_msg.h"
 #include "pfile.h"
 #include "plrmsg.h"
 #include "utils/sdl_geometry.h"
