+++ Source/inv.cpp
@@ -21,7 +21,6 @@
 #include "inv_iterators.hpp"
 #include "levels/town.h"
 #include "minitext.h"
-#include "miniwin/misc_msg.h"
 #include "options.h"
 #include "panels/ui_panels.hpp"
 #include "plrmsg.h"
