+++ Source/qol/stash.cpp
@@ -16,7 +16,6 @@
 #include "engine/size.hpp"
 #include "hwcursor.hpp"
 #include "minitext.h"
-#include "miniwin/misc_msg.h"
 #include "stores.h"
 #include "utils/format_int.hpp"
 #include "utils/language.h"
